requirements.txt
pyodbc
selenium
lxml
names
python-dotenv